chore: menu refactoring
